+++ x/leverage/client/tests/tests.go
@@ -95,10 +95,11 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 			false,
 			&types.QueryMarketSummaryResponse{},
 			&types.QueryMarketSummaryResponse{
-				SymbolDenom:        "UMEE",
-				Exponent:           6,
-				OraclePrice:        &oracleSymbolPrice,
-				UTokenExchangeRate: sdk.OneDec(),
+				SymbolDenom:         "UMEE",
+				Exponent:            6,
+				OraclePrice:         &oracleSymbolPrice,
+				OracleHistoricPrice: &oracleSymbolPrice,
+				UTokenExchangeRate:  sdk.OneDec(),
 				// Borrow rate * (1.52 - ReserveFactor - OracleRewardFactor)
 				// 1.52 * (1 - 0.2 - 0.01) = 1.2008
 				Supply_APY: sdk.MustNewDecFromStr("1.2008"),
@@ -292,6 +293,12 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 				BorrowedValue: sdk.MustNewDecFromStr("0.00858671"),
 				// (1001 / 1000000) * 34.21 * 0.25 = 0.0085610525
 				BorrowLimit: sdk.MustNewDecFromStr("0.0085610525"),
+				// (251 / 1000000) * 34.21 = 0.00858671
+				HistoricBorrowedValue: sdk.MustNewDecFromStr("0.00858671"),
+				// (1001 / 1000000) * 34.21 * 0.25 = 0.0085610525
+				HistoricBorrowLimit: sdk.MustNewDecFromStr("0.0085610525"),
+				// (1001 / 1000000) * 34.21 * 0.25 = 0.0085610525
+				SpotBorrowLimit: sdk.MustNewDecFromStr("0.0085610525"),
 				// (1001 / 1000000) * 0.25 * 34.21 = 0.0085610525
 				LiquidationThreshold: sdk.MustNewDecFromStr("0.0085610525"),
 			},