+++ x/oracle/keeper/grpc_query_test.go
@@ -210,27 +210,50 @@ func (s *IntegrationTestSuite) TestQuerier_Medians() {
 	atomMedian1 := sdk.DecCoin{Denom: "atom", Amount: sdk.MustNewDecFromStr("51.09")}
 	umeeMedian1 := sdk.DecCoin{Denom: "umee", Amount: sdk.MustNewDecFromStr("6540.23")}
 
-	app.OracleKeeper.SetHistoricMedian(ctx, atomMedian0.Denom, uint64(ctx.BlockHeight()-4), atomMedian0.Amount)
-	app.OracleKeeper.SetHistoricMedian(ctx, umeeMedian0.Denom, uint64(ctx.BlockHeight()-4), umeeMedian0.Amount)
+	blockHeight0 := uint64(ctx.BlockHeight() - 4)
+	app.OracleKeeper.SetHistoricMedian(ctx, atomMedian0.Denom, blockHeight0, atomMedian0.Amount)
+	app.OracleKeeper.SetHistoricMedian(ctx, umeeMedian0.Denom, blockHeight0, umeeMedian0.Amount)
 
 	res, err := s.queryClient.Medians(ctx.Context(), &types.QueryMedians{})
 	s.Require().NoError(err)
-	s.Require().Equal(res.Medians, sdk.NewDecCoins(atomMedian0, umeeMedian0))
+
+	expected := []types.Price{
+		*types.NewPrice(atomMedian0.Amount, "atom", blockHeight0),
+		*types.NewPrice(umeeMedian0.Amount, "umee", blockHeight0),
+	}
+	s.Require().Equal(res.Medians, expected)
 
 	res, err = s.queryClient.Medians(ctx.Context(), &types.QueryMedians{Denom: atomMedian0.Denom, NumStamps: 1})
 	s.Require().NoError(err)
-	s.Require().Equal(res.Medians, sdk.NewDecCoins(atomMedian0))
 
-	app.OracleKeeper.SetHistoricMedian(ctx, atomMedian1.Denom, uint64(ctx.BlockHeight()-2), atomMedian1.Amount)
-	app.OracleKeeper.SetHistoricMedian(ctx, umeeMedian1.Denom, uint64(ctx.BlockHeight()-2), umeeMedian1.Amount)
+	expected = []types.Price{
+		*types.NewPrice(atomMedian0.Amount, "atom", blockHeight0),
+	}
+	s.Require().Equal(res.Medians, expected)
+
+	blockHeight1 := uint64(ctx.BlockHeight() - 2)
+	app.OracleKeeper.SetHistoricMedian(ctx, atomMedian1.Denom, blockHeight1, atomMedian1.Amount)
+	app.OracleKeeper.SetHistoricMedian(ctx, umeeMedian1.Denom, blockHeight1, umeeMedian1.Amount)
 
 	res, err = s.queryClient.Medians(ctx.Context(), &types.QueryMedians{})
 	s.Require().NoError(err)
-	s.Require().Equal(res.Medians, sdk.DecCoins{atomMedian0, atomMedian1, umeeMedian0, umeeMedian1})
+
+	expected = []types.Price{
+		*types.NewPrice(atomMedian0.Amount, "atom", blockHeight0),
+		*types.NewPrice(umeeMedian0.Amount, "umee", blockHeight0),
+		*types.NewPrice(atomMedian1.Amount, "atom", blockHeight1),
+		*types.NewPrice(umeeMedian1.Amount, "umee", blockHeight1),
+	}
+	s.Require().Equal(res.Medians, expected)
 
 	res, err = s.queryClient.Medians(ctx.Context(), &types.QueryMedians{Denom: atomMedian1.Denom, NumStamps: 2})
 	s.Require().NoError(err)
-	s.Require().Equal(res.Medians, sdk.DecCoins{atomMedian1, atomMedian0})
+
+	expected = []types.Price{
+		*types.NewPrice(atomMedian0.Amount, "atom", blockHeight0),
+		*types.NewPrice(atomMedian1.Amount, "atom", blockHeight1),
+	}
+	s.Require().Equal(res.Medians, expected)
 
 	res, err = s.queryClient.Medians(ctx.Context(), &types.QueryMedians{Denom: atomMedian1.Denom, NumStamps: 0})
 	s.Require().ErrorContains(err, "parameter NumStamps must be greater than 0")
@@ -244,16 +267,26 @@ func (s *IntegrationTestSuite) TestQuerier_MedianDeviations() {
 	umeeMedianDeviation := sdk.DecCoin{Denom: "umee", Amount: sdk.MustNewDecFromStr("9541.48")}
 
 	app.OracleKeeper.SetMedianStampPeriod(ctx, 1)
-	app.OracleKeeper.SetHistoricMedianDeviation(ctx, atomMedianDeviation.Denom, uint64(ctx.BlockHeight()-1), atomMedianDeviation.Amount)
-	app.OracleKeeper.SetHistoricMedianDeviation(ctx, umeeMedianDeviation.Denom, uint64(ctx.BlockHeight()-1), umeeMedianDeviation.Amount)
+	blockHeight := uint64(ctx.BlockHeight() - 1)
+	app.OracleKeeper.SetHistoricMedianDeviation(ctx, atomMedianDeviation.Denom, blockHeight, atomMedianDeviation.Amount)
+	app.OracleKeeper.SetHistoricMedianDeviation(ctx, umeeMedianDeviation.Denom, blockHeight, umeeMedianDeviation.Amount)
 
 	res, err := s.queryClient.MedianDeviations(ctx.Context(), &types.QueryMedianDeviations{})
 	s.Require().NoError(err)
-	s.Require().Equal(res.MedianDeviations, sdk.NewDecCoins(atomMedianDeviation, umeeMedianDeviation))
+
+	expected := []types.Price{
+		*types.NewPrice(atomMedianDeviation.Amount, "atom", blockHeight),
+		*types.NewPrice(umeeMedianDeviation.Amount, "umee", blockHeight),
+	}
+	s.Require().Equal(res.MedianDeviations, expected)
 
 	res, err = s.queryClient.MedianDeviations(ctx.Context(), &types.QueryMedianDeviations{Denom: atomMedianDeviation.Denom})
 	s.Require().NoError(err)
-	s.Require().Equal(res.MedianDeviations, sdk.NewDecCoins(atomMedianDeviation))
+
+	expected = []types.Price{
+		*types.NewPrice(atomMedianDeviation.Amount, "atom", blockHeight),
+	}
+	s.Require().Equal(res.MedianDeviations, expected)
 }
 
 func (s *IntegrationTestSuite) TestEmptyRequest() {